+++ legwork/tests/test_utils.py
@@ -25,27 +25,6 @@ class Test(unittest.TestCase):
 
         self.assertTrue(np.allclose(f_orb, should_be_f_orb))
 
-    def test_bad_input(self):
-        """check functions can deal with bad input"""
-        n_vals = 10000
-        f_orb_i = 10**(np.random.uniform(-5, -1, n_vals)) * u.Hz
-        ecc_i = np.zeros(n_vals)
-        m_1 = np.random.uniform(0, 50, n_vals) * u.Msun
-        m_2 = np.random.uniform(0, 50, n_vals) * u.Msun
-
-        # check the function doesn't crash if you don't give the chirp mass
-        utils.determine_stationarity(f_orb_i=f_orb_i, t_evol=4 * u.yr,
-                                     ecc_i=ecc_i, m_1=m_1, m_2=m_2)
-
-        # check that it *does* crash when no masses supplied
-        no_worries = True
-        try:
-            utils.determine_stationarity(f_orb_i=f_orb_i, t_evol=4 * u.yr,
-                                         ecc_i=ecc_i)
-        except ValueError:
-            no_worries = False
-        self.assertFalse(no_worries)
-
     def test_average_response(self):
         """make sure that response integrals are correct"""
         # based on Babak+2021, the sum of the average responses <F_plus^2 a_plus^2>
@@ -64,9 +43,9 @@ class Test(unittest.TestCase):
 
         result, error = integrate.nquad(
             integrand,
-            [[0, np.pi],      # theta
-             [0, 2 * np.pi],  # phi
-             [0, 2 * np.pi]]) # psi
+            [[0, np.pi],       # theta
+             [0, 2 * np.pi],   # phi
+             [0, 2 * np.pi]])  # psi
 
         self.assertAlmostEqual(result, 3/10)
 
@@ -89,6 +68,3 @@ class Test(unittest.TestCase):
              [0, np.pi / 2]])  # inc
 
         self.assertAlmostEqual(result, 0.12)
-
-
-