+++ legwork/visualisation.py
@@ -311,7 +311,7 @@ def plot_sensitivity_curve(frequency_range=None, y_quantity="ASD", fig=None, ax=
     # plot the curve and fill if needed
     ax.loglog(frequency_range, noise_amplitude, color=color, label=label, linewidth=linewidth)
     if fill:
-        ax.fill_between(frequency_range, 0, noise_amplitude, alpha=alpha, color=color)
+        ax.fill_between(frequency_range, 0 * noise_amplitude.unit, noise_amplitude, alpha=alpha, color=color)
 
     # adjust labels, sizes and frequency limits to plot is flush to the edges
     ax.set_xlabel(r'Frequency [$\rm Hz$]')
@@ -329,9 +329,9 @@ def plot_sensitivity_curve(frequency_range=None, y_quantity="ASD", fig=None, ax=
     return fig, ax
 
 
-def plot_sources_on_sc_circ_stat(f_orb, h_0_2, snr, weights=None, snr_cutoff=0, t_obs=4 * u.yr,
-                                 instrument="LISA", custom_psd=None, L=2.5e9 * u.m, approximate_R=False,
-                                 confusion_noise=True, fig=None, ax=None, show=True, **kwargs):
+def plot_sources_on_sc_circ_stat(f_orb, h_0_2, snr, weights=None, snr_cutoff=0, t_obs="auto",
+                                 instrument="LISA", custom_psd=None, L="auto", approximate_R=False,
+                                 confusion_noise="auto", fig=None, ax=None, show=True, **kwargs):
     """Overlay circular/stationary sources on the LISA sensitivity curve.
 
     Each source is plotted at its gravitational wave frequency (n = 2) such that its height above the curve
@@ -362,18 +362,20 @@ def plot_sources_on_sc_circ_stat(f_orb, h_0_2, snr, weights=None, snr_cutoff=0, 
         even if it ignores some.
 
     t_obs : `float`
-        Observation time (default 4 years)
+        Observation time (default "auto")
 
     L : `float`
-        LISA arm length in metres
+        Arm length
 
     approximate_R : `boolean`
         Whether to approximate the response function (default: no)
 
-    confusion_noise  : `various`
-        Galactic confusion noise. Acceptable inputs are 'robson19' (the confusion noise from Robson+19),
-        `None` (don't include confusion noise) or a custom function that gives the confusion noise at each
-        frequency for a given mission length where it would be called by running `noise(f, t_obs)`
+    confusion_noise : `various`
+        Galactic confusion noise. Acceptable inputs are either one of the values listed in
+        :meth:`legwork.psd.get_confusion_noise`, "auto" (automatically selects confusion noise based on
+        `instrument` - 'robson19' if LISA and 'huang20' if TianQin), or a custom function that gives the
+        confusion noise at each frequency for a given mission length where it would be called by running
+        `noise(f, t_obs)` and return a value with units of inverse Hertz
 
     fig: `matplotlib Figure`
         A figure on which to plot the distribution. Both `ax` and `fig` must be supplied for either to be used
@@ -409,6 +411,10 @@ def plot_sources_on_sc_circ_stat(f_orb, h_0_2, snr, weights=None, snr_cutoff=0, 
         print("ERROR: There are no binaries above provided `snr_cutoff`")
         return fig, ax
 
+    # convert auto observation times
+    if t_obs == "auto":
+        t_obs = 4 * u.yr if instrument == "LISA" else 5 * u.yr
+
     # calculate the GW frequency and ASD for detectable binaries
     f_GW = f_orb[detectable] * 2
     asd = ((1/4 * t_obs)**(1/2) * h_0_2[detectable]).to(u.Hz**(-1/2))
@@ -425,9 +431,9 @@ def plot_sources_on_sc_circ_stat(f_orb, h_0_2, snr, weights=None, snr_cutoff=0, 
     return fig, ax
 
 
-def plot_sources_on_sc_ecc_stat(f_dom, snr, weights=None, snr_cutoff=0, t_obs=4 * u.yr,
-                                instrument="LISA", custom_psd=None, L=2.5e9 * u.m, approximate_R=False,
-                                confusion_noise=True, fig=None, ax=None, show=True, **kwargs):
+def plot_sources_on_sc_ecc_stat(f_dom, snr, weights=None, snr_cutoff=0, t_obs="auto",
+                                instrument="LISA", custom_psd=None, L="auto", approximate_R=False,
+                                confusion_noise="auto", fig=None, ax=None, show=True, **kwargs):
     """Overlay eccentric/stationary sources on the LISA sensitivity curve.
 
     Each source is plotted at its max snr harmonic frequency such that that its height above the curve is
@@ -455,18 +461,20 @@ def plot_sources_on_sc_ecc_stat(f_dom, snr, weights=None, snr_cutoff=0, t_obs=4 
         even if it ignores some.
 
     t_obs : `float`
-        Observation time (default 4 years)
+        Observation time (default auto)
 
     L : `float`
-        LISA arm length in metres
+        Arm length in metres
 
     approximate_R : `boolean`
         Whether to approximate the response function (default: no)
 
-    confusion_noise  : `various`
-        Galactic confusion noise. Acceptable inputs are 'robson19' (the confusion noise from Robson+19),
-        `None` (don't include confusion noise) or a custom function that gives the confusion noise at each
-        frequency for a given mission length where it would be called by running `noise(f, t_obs)`
+    confusion_noise : `various`
+        Galactic confusion noise. Acceptable inputs are either one of the values listed in
+        :meth:`legwork.psd.get_confusion_noise`, "auto" (automatically selects confusion noise based on
+        `instrument` - 'robson19' if LISA and 'huang20' if TianQin), or a custom function that gives the
+        confusion noise at each frequency for a given mission length where it would be called by running
+        `noise(f, t_obs)` and return a value with units of inverse Hertz
 
     fig: `matplotlib Figure`
         A figure on which to plot the distribution. Both `ax` and `fig` must be supplied for either to be used